ios: do not consider connected until receiver app has namespaces registered
The chromecast allows a second device to connect while the receiver
app is still booting. This device gets an didConnectToApplication
callback before the receiver app has registered a namespace, and thus
the earliest messages it sends can be missed.

We work around this bug but not considering the connection to be
successful until the reciever app has set its namespaces.  (The
connected sender is notified by an updated GCKApplicationMetadata
status object.)